requirements.txt
sentence-transformers>=2.6.0
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-huggingface
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental